+++ internal/app/wwctl/profile/list/main.go
@@ -17,11 +17,10 @@ func CobraRunE(vars *variables) func(cmd *cobra.Command, args []string) (err err
 			args = strings.FieldsFunc(args[0], func(r rune) bool { return r == ',' })
 		}
 		req := wwapiv1.GetProfileList{
-			ShowAll:     vars.showAll,
-			ShowFullAll: vars.showFullAll,
-			ShowYaml:    vars.showYaml,
-			ShowJson:    vars.showJson,
-			Profiles:    args,
+			ShowAll:  vars.showAll,
+			ShowYaml: vars.showYaml,
+			ShowJson: vars.showJson,
+			Profiles: args,
 		}
 		profileInfo, err := apiprofile.ProfileList(&req)
 		if err != nil {