+++ internal/app/wwctl/profile/list/root.go
@@ -3,10 +3,9 @@ package list
 import "github.com/spf13/cobra"
 
 type variables struct {
-	showAll     bool
-	showFullAll bool
-	showYaml    bool
-	showJson    bool
+	showAll  bool
+	showYaml bool
+	showJson bool
 }
 
 // GetRootCommand returns the root cobra.Command for the application.
@@ -21,7 +20,6 @@ func GetCommand() *cobra.Command {
 		Aliases:               []string{"ls"},
 	}
 	baseCmd.PersistentFlags().BoolVarP(&vars.showAll, "all", "a", false, "Show all profile configurations")
-	baseCmd.PersistentFlags().BoolVarP(&vars.showFullAll, "fullall", "A", false, "Show all profile configurations inclusive empty entries")
 	baseCmd.PersistentFlags().BoolVarP(&vars.showYaml, "yaml", "y", false, "Show profile configurations via yaml format")
 	baseCmd.PersistentFlags().BoolVarP(&vars.showJson, "json", "j", false, "Show profile configurations via json format")